\docType{package}
\name{cdtRprogramming-package}
\alias{cdtRprogramming}
\alias{cdtRprogramming-package}
\title{R courses for Newcastle University's CDT in Cloud Computing for Big Data}
\description{
  Functions and datasets used for Newcastle University's 
  CDT in Cloud Computing for Big Data. This package is 
  used in the "R Programming" course.
}
\keyword{package}

